first fix with Double
Will be edited with other Number

+++ src/main/java/calculator/Minus.java
@@ -42,11 +42,12 @@ public final class Minus extends Operation
 
     /**
      * The actual computation of the (binary) arithmetic subtraction of two integers
+     *
      * @param l The first integer
      * @param r The second integer that should be subtracted from the first
      * @return The integer that is the result of the subtraction
      */
-  public int op(int l, int r) {
+  public double op(double l, double r) {
   	return (l-r);
   }
 }
